requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.7
h11==0.14.0
idna==3.10
numpy==2.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
pyserial==3.5
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.34.0
